refactor: split utils into logging, streaming, and demo launchers
