+++ research-agent/utils/__init__.py
@@ -1,12 +1,20 @@
-"""
-Utilities package for Deep Research Agency Tutorial.
+"""Utilities package for Deep Research Agency Tutorial.
 
 This package contains:
-- demo: Demo and UI utilities for running agencies
+- logging: Debug logging helpers
+- streaming: Streaming logic for demos
+- demo_launchers: Demo and UI utilities for running agencies
 - pdf: PDF generation utilities for research reports
 """
 
-from .demo import copilot_demo, stream_demo, run_agency_demo
+from .demo_launchers import copilot_demo, run_agency_demo, save_research_report
+from .streaming import stream_demo
 from .pdf import save_research_to_pdf
 
-__all__ = ["copilot_demo", "stream_demo", "run_agency_demo", "save_research_to_pdf"]
+__all__ = [
+    "copilot_demo",
+    "run_agency_demo",
+    "save_research_report",
+    "stream_demo",
+    "save_research_to_pdf",
+]
